+++ lib/pages/home_page.dart
@@ -1,32 +1,49 @@
 import 'package:flutter/material.dart';
-import 'package:carousel_slider/carousel_slider.dart';
-import 'package:glass/glass.dart';
+import 'package:helpus/Widgets/custom_shalter_card.dart';
 import 'package:helpus/pages/addhome.dart';
+import 'package:helpus/widgets/custom_category_card.dart';
+import 'package:helpus/widgets/custom_image_carousel.dart';
+import 'package:helpus/widgets/custom_shelter_card.dart';
 
 class HomePage extends StatelessWidget {
   const HomePage({super.key});
 
   @override
   Widget build(BuildContext context) {
+    Size size = MediaQuery.of(context).size;
+
     return Scaffold(
       appBar: AppBar(
-        title: const SizedBox(
-            width: 100,
-            child: Image(
-              image: AssetImage('assets/images/helpus.png'),
-              fit: BoxFit.cover,
-            )),
+        forceMaterialTransparency: true,
+        automaticallyImplyLeading: false,
+        title: const Padding(
+          padding: EdgeInsets.only(left: 10.0),
+          child: SizedBox(
+              width: 100,
+              child: Image(
+                image: AssetImage('assets/images/helpus.png'),
+                fit: BoxFit.cover,
+              )),
+        ),
         actions: [
-          IconButton(
-            onPressed: () {},
-            icon: const Icon(Icons.notifications, color: Colors.black),
+          Padding(
+            padding: const EdgeInsets.only(right: 10.0),
+            child: IconButton(
+              iconSize: 30,
+              onPressed: () {},
+              icon: const Icon(Icons.notifications, color: Colors.black),
+            ),
           )
         ],
       ),
       body: SingleChildScrollView(
+        physics: const BouncingScrollPhysics(),
         child: Column(
           mainAxisAlignment: MainAxisAlignment.start,
           children: [
+            SizedBox(
+              height: size.height * 0.01,
+            ),
             CarouselWithDotsPage(imgList: const [
               'assets/images/1.jpg',
               'assets/images/2.jpg',
@@ -43,8 +60,8 @@ class HomePage extends StatelessWidget {
                     title: 'Education', image: 'assets/images/cat-4.png'),
               ],
             ),
-            const SizedBox(
-              height: 30,
+            SizedBox(
+              height: size.height * 0.03,
             ),
             const Padding(
               padding: EdgeInsets.symmetric(horizontal: 20),
@@ -60,34 +77,33 @@ class HomePage extends StatelessWidget {
                 ),
               ),
             ),
-            const SizedBox(
-              height: 20,
-            ),
+            SizedBox(height: size.height * 0.02),
             SizedBox(
               height: 180,
               child: ListView(
+                physics: const BouncingScrollPhysics(),
                 scrollDirection: Axis.horizontal,
-                children: const [
+                children: [
                   SizedBox(
-                    width: 20,
+                    width: size.width * 0.05,
                   ),
-                  NeedFirstBox(
+                  const NeedFirstBox(
                     title: 'Shakti NGO Shelter Home',
                     image: 'assets/images/ngo-banner-1.png',
                   ),
-                  NeedFirstBox(
+                  const NeedFirstBox(
                     title: 'Homies NGO Shelter Home',
                     image: 'assets/images/ngo-banner-2.png',
                   ),
-                  NeedFirstBox(
+                  const NeedFirstBox(
                     title: 'UNEX NGO Shelter Home',
                     image: 'assets/images/ngo-banner-3.png',
                   ),
                 ],
               ),
             ),
-            const SizedBox(
-              height: 20,
+            SizedBox(
+              height: size.height * 0.03,
             ),
             const Padding(
               padding: EdgeInsets.symmetric(horizontal: 20),
@@ -103,8 +119,8 @@ class HomePage extends StatelessWidget {
                 ),
               ),
             ),
-            const SizedBox(
-              height: 20,
+            SizedBox(
+              height: size.height * 0.02,
             ),
             const LatestShalter(
               title: 'Shakti NGO Shelter Home',
@@ -128,257 +144,11 @@ class HomePage extends StatelessWidget {
         backgroundColor: const Color(0xffFEB61D),
         onPressed: () {
           Navigator.push(context, MaterialPageRoute(builder: (context) {
-            return AddHomeShelter();
+            return const AddHomeShelter();
           }));
         },
         child: const Icon(Icons.add),
       ),
-    );
-  }
-}
-
-class LatestShalter extends StatelessWidget {
-  const LatestShalter({
-    super.key,
-    required this.title,
-    required this.image,
-    required this.distance,
-  });
-  final String title;
-  final String image;
-  final String distance;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-      child: Row(
-        children: [
-          Container(
-            height: 80,
-            width: 80,
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(20),
-              image: DecorationImage(
-                image: AssetImage(image),
-                fit: BoxFit.cover,
-              ),
-              color: Colors.red,
-            ),
-          ),
-          const SizedBox(
-            width: 20,
-          ),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                title,
-                style: const TextStyle(
-                  fontSize: 18,
-                  fontWeight: FontWeight.w600,
-                  color: Colors.black,
-                ),
-                overflow: TextOverflow.fade,
-              ),
-              Text(
-                distance,
-                style: const TextStyle(
-                  fontSize: 14,
-                  fontWeight: FontWeight.w600,
-                  color: Colors.black,
-                ),
-              ),
-            ],
-          )
-        ],
-      ),
-    );
-  }
-}
-
-class NeedFirstBox extends StatelessWidget {
-  const NeedFirstBox({
-    super.key,
-    required this.title,
-    required this.image,
-  });
-  final String title;
-  final String image;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(right: 20),
-      width: 300,
-      height: 180,
-      decoration: BoxDecoration(
-        image: DecorationImage(
-          image: AssetImage(image),
-          fit: BoxFit.cover,
-        ),
-        borderRadius: BorderRadius.circular(10),
-        //
-      ),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.end,
-        children: [
-          Container(
-            height: 50,
-            width: double.infinity,
-            decoration: const BoxDecoration(
-              borderRadius: BorderRadius.only(
-                bottomLeft: Radius.circular(10),
-                bottomRight: Radius.circular(10),
-              ),
-            ),
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-              child: Text(
-                title,
-                style: const TextStyle(
-                    overflow: TextOverflow.ellipsis,
-                    color: Colors.white,
-                    fontSize: 18,
-                    fontWeight: FontWeight.bold),
-              ),
-            ),
-          ).asGlass(
-            tintColor: Colors.black.withOpacity(0.2),
-            clipBorderRadius: const BorderRadius.only(
-              bottomLeft: Radius.circular(10),
-              bottomRight: Radius.circular(10),
-            ),
-          )
-        ],
-      ),
-    );
-  }
-}
-
-class HomeCategory extends StatelessWidget {
-  const HomeCategory({
-    super.key,
-    required this.title,
-    required this.image,
-  });
-  final String title;
-  final String image;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Container(
-          margin: const EdgeInsets.all(10),
-          width: 60,
-          height: 60,
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(500),
-            image: DecorationImage(
-              image: AssetImage(image),
-              fit: BoxFit.cover,
-            ),
-          ),
-        ),
-        Text(
-          title,
-          style: const TextStyle(
-            fontSize: 16,
-            fontWeight: FontWeight.w600,
-            color: Colors.black,
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-// ignore: must_be_immutable
-class CarouselWithDotsPage extends StatefulWidget {
-  List<String> imgList;
-
-  // ignore: use_key_in_widget_constructors
-  CarouselWithDotsPage({required this.imgList});
-
-  @override
-  // ignore: library_private_types_in_public_api
-  _CarouselWithDotsPageState createState() => _CarouselWithDotsPageState();
-}
-
-class _CarouselWithDotsPageState extends State<CarouselWithDotsPage> {
-  int _current = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    final List<Widget> imageSliders = widget.imgList
-        .map(
-          (item) => ClipRRect(
-            borderRadius: const BorderRadius.all(
-              Radius.circular(5.0),
-            ),
-            child: Stack(
-              children: [
-                Image.asset(
-                  item,
-                  fit: BoxFit.cover,
-                  width: double.infinity,
-                ),
-                Positioned(
-                  bottom: 0.0,
-                  left: 0.0,
-                  right: 0.0,
-                  child: Container(
-                    padding: const EdgeInsets.symmetric(
-                      horizontal: 20,
-                      vertical: 10,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        )
-        .toList();
-
-    return Column(
-      children: [
-        const Padding(
-          padding: EdgeInsets.all(5),
-        ),
-        CarouselSlider(
-          items: imageSliders,
-          options: CarouselOptions(
-              autoPlay: true,
-              enlargeCenterPage: true,
-              aspectRatio: 2.2,
-              onPageChanged: (index, reason) {
-                setState(() {
-                  _current = index;
-                });
-              }),
-        ),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: widget.imgList.map((url) {
-            int index = widget.imgList.indexOf(url);
-            return Container(
-              width: 8,
-              height: 8,
-              margin: const EdgeInsets.symmetric(
-                vertical: 10,
-                horizontal: 3,
-              ),
-              decoration: BoxDecoration(
-                shape: BoxShape.circle,
-                color: _current == index
-                    ? const Color.fromRGBO(0, 0, 0, 0.9)
-                    : const Color.fromRGBO(0, 0, 0, 0.4),
-              ),
-            );
-          }).toList(),
-        )
-      ],
     );
   }
 }
