+++ lib/widgets/custom_category_card.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/material.dart';
+
+class HomeCategory extends StatelessWidget {
+  const HomeCategory({
+    super.key,
+    required this.title,
+    required this.image,
+  });
+  final String title;
+  final String image;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        Container(
+          margin: const EdgeInsets.all(10),
+          width: 60,
+          height: 60,
+          decoration: BoxDecoration(
+            borderRadius: BorderRadius.circular(500),
+            image: DecorationImage(
+              image: AssetImage(image),
+              fit: BoxFit.cover,
+            ),
+          ),
+        ),
+        Text(
+          title,
+          style: const TextStyle(
+            fontSize: 16,
+            fontWeight: FontWeight.w600,
+            color: Colors.black,
+          ),
+        ),
+      ],
+    );
+  }
+}
