+++ lib/widgets/custom_image_carousel.dart
@@ -1,0 +1,72 @@
+import 'package:carousel_slider/carousel_slider.dart';
+import 'package:flutter/material.dart';
+
+class CarouselWithDotsPage extends StatefulWidget {
+  List<String> imgList;
+
+  // ignore: use_key_in_widget_constructors
+  CarouselWithDotsPage({required this.imgList});
+
+  @override
+  // ignore: library_private_types_in_public_api
+  _CarouselWithDotsPageState createState() => _CarouselWithDotsPageState();
+}
+
+class _CarouselWithDotsPageState extends State<CarouselWithDotsPage> {
+  int _current = 0;
+
+  @override
+  Widget build(BuildContext context) {
+    final List<Widget> imageSliders = widget.imgList
+        .map(
+          (item) => ClipRRect(
+            borderRadius: const BorderRadius.all(
+              Radius.circular(5.0),
+            ),
+            child: Image.asset(
+              item,
+              fit: BoxFit.cover,
+              width: double.infinity,
+            ),
+          ),
+        )
+        .toList();
+
+    return Column(
+      children: [
+        CarouselSlider(
+          items: imageSliders,
+          options: CarouselOptions(
+              autoPlay: true,
+              enlargeCenterPage: true,
+              aspectRatio: 2.2,
+              onPageChanged: (index, reason) {
+                setState(() {
+                  _current = index;
+                });
+              }),
+        ),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: widget.imgList.map((url) {
+            int index = widget.imgList.indexOf(url);
+            return Container(
+              width: 8,
+              height: 8,
+              margin: const EdgeInsets.symmetric(
+                vertical: 10,
+                horizontal: 3,
+              ),
+              decoration: BoxDecoration(
+                shape: BoxShape.circle,
+                color: _current == index
+                    ? const Color.fromRGBO(0, 0, 0, 0.9)
+                    : const Color.fromRGBO(0, 0, 0, 0.4),
+              ),
+            );
+          }).toList(),
+        )
+      ],
+    );
+  }
+}
