+++ lib/widgets/custom_shalter_card.dart
@@ -1,0 +1,61 @@
+import 'package:flutter/material.dart';
+
+class LatestShalter extends StatelessWidget {
+  const LatestShalter({
+    super.key,
+    required this.title,
+    required this.image,
+    required this.distance,
+  });
+  final String title;
+  final String image;
+  final String distance;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
+      child: Row(
+        children: [
+          Container(
+            height: 80,
+            width: 80,
+            decoration: BoxDecoration(
+              borderRadius: BorderRadius.circular(20),
+              image: DecorationImage(
+                image: AssetImage(image),
+                fit: BoxFit.cover,
+              ),
+              color: Colors.red,
+            ),
+          ),
+          const SizedBox(
+            width: 20,
+          ),
+          Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                title,
+                style: const TextStyle(
+                  fontSize: 18,
+                  fontWeight: FontWeight.w600,
+                  color: Colors.black,
+                ),
+                overflow: TextOverflow.fade,
+              ),
+              Text(
+                distance,
+                style: const TextStyle(
+                  fontSize: 14,
+                  fontWeight: FontWeight.w600,
+                  color: Colors.black,
+                ),
+              ),
+            ],
+          )
+        ],
+      ),
+    );
+  }
+}
