+++ lib/widgets/custom_shelter_card.dart
@@ -1,0 +1,60 @@
+import 'package:flutter/material.dart';
+import 'package:glass/glass.dart';
+
+class NeedFirstBox extends StatelessWidget {
+  const NeedFirstBox({
+    super.key,
+    required this.title,
+    required this.image,
+  });
+  final String title;
+  final String image;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.only(right: 20),
+      width: 300,
+      decoration: BoxDecoration(
+        image: DecorationImage(
+          image: AssetImage(image),
+          fit: BoxFit.cover,
+        ),
+        borderRadius: BorderRadius.circular(10),
+        //
+      ),
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.end,
+        children: [
+          Container(
+            height: 50,
+            width: double.infinity,
+            decoration: const BoxDecoration(
+              borderRadius: BorderRadius.only(
+                bottomLeft: Radius.circular(10),
+                bottomRight: Radius.circular(10),
+              ),
+            ),
+            child: Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
+              child: Text(
+                title,
+                style: const TextStyle(
+                    overflow: TextOverflow.ellipsis,
+                    color: Colors.white,
+                    fontSize: 18,
+                    fontWeight: FontWeight.bold),
+              ),
+            ),
+          ).asGlass(
+            tintColor: Colors.black.withOpacity(0.2),
+            clipBorderRadius: const BorderRadius.only(
+              bottomLeft: Radius.circular(10),
+              bottomRight: Radius.circular(10),
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
